Add dummy implementations of some extension points
In order to get a feeling for order of execution and what all
these different extension points are for exactly, I've created
this branch. Note that as long as SimpleHighlightVisitor is
enabled in the plugin.xml, SimpleAnnotator will not get called,
because the platform's DefaultHighlightVisitor doesn't get called.

So in order to see SimpleAnnotator do its thing, comment out
SimpleHighlightVisitor in plugin.xml.